doc: readme added
and yes, that was a big rant, and yes that last commit was huge instead
of incremental, but I was trying to wait till I actually got something
to work correctly to commit something, not just once I copied someone
else's code and changed a few variables to get it to work with what I
already had. but oh well, after accidently breaking the project and
working for 2+ hours to find out what the 98ser8mt*#&psurgu was going on
this is all I've done so far and hopefully I can make more progress
tomorrow.